Reworked

+++ Zadanie24/Program.cs
@@ -1,7 +1,7 @@
 ﻿// Найти кубы чисел от 1 до N
 int a = 1;
 Console.WriteLine("Введите конечное число ");
-double b = Convert.ToDouble(Console.ReadLine());
+int b = Convert.ToInt32(Console.ReadLine());
               while (a <= b)
              {
                  Console.WriteLine("Куб числа " + a);
